+++ concordium-sdk/src/main/java/com/concordium/sdk/ClientV2MapperExtensions.java
@@ -29,6 +29,8 @@ import com.concordium.sdk.crypto.ed25519.ED25519PublicKey;
 import com.concordium.sdk.crypto.elgamal.ElgamalPublicKey;
 import com.concordium.sdk.crypto.pedersencommitment.PedersenCommitmentKey;
 import com.concordium.sdk.crypto.pointchevalsanders.PSPublicKey;
+import com.concordium.sdk.responses.BlockIdentifier;
+import com.concordium.sdk.responses.KeyValurPair;
 import com.concordium.sdk.requests.AccountQuery;
 import com.concordium.sdk.requests.BlockQuery;
 import com.concordium.sdk.responses.Epoch;
@@ -916,13 +918,6 @@ interface ClientV2MapperExtensions {
 
     static com.concordium.sdk.types.ContractAddress to(ContractAddress address) {
         return com.concordium.sdk.types.ContractAddress.from(address.getIndex(), address.getSubindex());
-    }
-
-    static ContractAddress to(com.concordium.sdk.types.ContractAddress contractAddress) {
-        return ContractAddress.newBuilder()
-                .setIndex(contractAddress.getIndex())
-                .setSubindex(contractAddress.getSubIndex())
-                .build();
     }
 
     static WasmModule to(VersionedModuleSource deployModule) {
@@ -2057,6 +2052,20 @@ interface ClientV2MapperExtensions {
         return VerificationKey.builder()
                 .verifyKey(k.getValue().toByteArray())
                 .signingScheme(SigningScheme.ED25519)
+                .build();
+    }
+
+    static ContractAddress to(com.concordium.sdk.types.ContractAddress contractAddress) {
+        return ContractAddress.newBuilder()
+                .setIndex(contractAddress.getIndex())
+                .setSubindex(contractAddress.getSubIndex())
+                .build();
+    }
+
+    static KeyValurPair to(InstanceStateKVPair instanceStateKVPair) {
+        return KeyValurPair.builder()
+                .key(instanceStateKVPair.getKey().toByteArray())
+                .value(instanceStateKVPair.getValue().toByteArray())
                 .build();
     }
 }